% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSimu1.R
\name{geneSimu1}
\alias{geneSimu1}
\title{Generate the data for simulation 1}
\usage{
geneSimu1(seed = 1)
}
\arguments{
\item{seed}{numeric number, random seed}
}
\value{
a data list with following items
	 Y: observation vector of length n;
   X: n by p covariate matrix;
   U1: p by m1 matrix, indicate feature membership
   types: the type of each feature, different type can have different s^2
  true_beta: true coefficients
}
\description{
Generate the data for simulation 1
}
