% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSimu2.R
\name{geneSimu2}
\alias{geneSimu2}
\title{Generate the data for simulation 2}
\usage{
geneSimu2(seed = 1)
}
\arguments{
\item{seed}{numeric number, random seed}
}
\value{
a data list with following items
	 Y: observation vector of length n;
   X: n by p covariate matrix;
   U1: p by m1 matrix, indicate feature membership
   types: the type of each feature, different type can have different s^2
   true_beta: true coefficients
}
\description{
Generate the data for simulation 2
}
