% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSimu4.R
\name{geneSimu4}
\alias{geneSimu4}
\title{Generate the data for simulation 4}
\usage{
geneSimu4(U = 0.2, seed = 1)
}
\arguments{
\item{U}{number, effect size}

\item{seed}{integer, specifying the random seed}
}
\value{
a data list with following items
  Y: observation vector of length n;
  X: n by p covariate matrix;
  U1: p by m1 matrix, indicate feature membership
  U2: m1 by m2 matrix, indicate level-1 group membership
  types: the type of each feature, different type can have different s^2
  true_beta: true coefficients
}
\description{
Generate the data for simulation 4
}
