% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSimu5.R
\name{geneSimu5}
\alias{geneSimu5}
\title{Generate the data for simulation 5}
\usage{
geneSimu5(N = 300, s2 = 5, m1_signal, m1_noise, p_perk = 20, a, b,
  seed = 123)
}
\arguments{
\item{N}{integer, sample size}

\item{s2}{numeric number, specifying s2 in SOG model}

\item{m1_signal}{integer, number of level-1 groups that are predictive of outcome}

\item{m1_noise}{integer, number of level-1 groups that are irrelevant to outcome}

\item{p_perk}{integer, number of features inside each group}

\item{a}{numeric number, a in beta distribution (prior of pi_k^0)}

\item{b}{numeric number, b in beta distribution (prior of pi_k^0)}

\item{seed}{integer, specifying the random seed}
}
\value{
a data list with following items
  Y: observation vector of length n;
  X: n by p covariate matrix;
  U1: p by m1 matrix, indicate feature membership
  types: the type of each feature, different type can have different s^2
  true_beta: true coefficients
  pi0_k: true pi_k^0
}
\description{
Generate the data for simulation 5
}
