\name{MOG-package}
\alias{MOG-package}
\alias{MOG}
\docType{package}
\title{
\packageTitle{MOG}
}
\description{
\packageDescription{MOG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MOG}
\packageIndices{MOG}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MOG}

Maintainer: \packageMaintainer{MOG}
}
\references{
~~ Literature or other references for background information ~~
}
