% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOG_binary.R
\name{MOG_binary}
\alias{MOG_binary}
\title{MOG function for binary outcome.}
\usage{
MOG_binary(X, Y, U1, U2, types = NULL, lassoInit = FALSE,
  weight_Tj = FALSE, weight_Dk = FALSE, pi2_prop_n = 10,
  pi3_prop_n = 10, MH_ind = 1, burnInIter = 1000, keepIter = 3000,
  maxIter = 10000, printInt = 100, seed = 123, alpha1 = 1,
  beta1 = 1, alpha2 = 1, beta2 = 1, alpha3 = 1, beta3 = 1)
}
\arguments{
\item{X}{n by p feature matrix, where n is the number of samples, 
p is number of features}

\item{Y}{outcome vector of length n}

\item{U1}{p by m1 matirx, denoting group membership of each feature}

\item{U2}{m1 by m2 matirx, denoting group membership of each level-1 group}

\item{types}{numeric vector of length p, type index of each feature, allows different variance 
in priors of coefficient of different types}

\item{lassoInit}{TRUE/FALSE, if TRUE, use lasso estimates as MCMC initials}

\item{weight_Tj}{TRUE/FALSE, if TRUE, the prior of gamma_jkl^0 will be weighted by T_j}

\item{weight_Dk}{TRUE/FALSE, if TRUE, the prior of gamma_kl^1 will be weighted by D_k}

\item{pi2_prop_n}{numeric number, specifying n_0 in the proposal distribution of pi_l^1}

\item{pi3_prop_n}{numeric number, specifying n_0 in the proposal distribution of pi_kl^0}

\item{MH_ind}{1/0, if 1 MH is used to update pi_k^0; otherwise, Gibbs sampling is used}

\item{burnInIter}{number of iterations as burn-in period}

\item{keepIter}{number of iterations to be stored}

\item{maxIter}{maximum number of iterations}

\item{printInt}{print progress message every print_int iterations}

\item{seed}{integer, specifying the random seed}

\item{alpha1}{numeric number, specifying alpha in the BETA distribution (prior of pi^2)}

\item{beta1}{numeric number, specifying beta in the BETA distribution (prior of pi^2)}

\item{alpha2}{numeric number, specifying alpha in the BETA distribution (prior of pi_l^1)}

\item{beta2}{numeric number, specifying beta in the BETA distribution (prior of pi_l^1)}

\item{alpha3}{numeric number, specifying alpha in the BETA distribution (prior of pi_kl^0)}

\item{beta3}{numeric number, specifying beta in the BETA distribution (prior of pi_kl^0)}
}
\value{
A list of MCMC output, together with feature selection probability 
and feature Bayesian qvalue
}
\description{
This function applies MOG to binary outcome
}
