% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOG_binary.R
\name{SOG_binary}
\alias{SOG_binary}
\title{SOG function for binary outcome}
\usage{
SOG_binary(X, Y, U1, types = NULL, lassoInit = FALSE,
  pi2_weighted = FALSE, pi2_prop_n = 10, MH_ind = 1,
  burnInIter = 1000, keepIter = 3000, maxIter = 10000,
  printInt = 1000, seed = 123, alpha2 = 1, beta2 = 1)
}
\arguments{
\item{X}{n by p feature matrix, where n is the number of samples, 
p is number of features}

\item{Y}{outcome vector of length n}

\item{U1}{p by m1 matirx, denoting group membership of each feature}

\item{types}{numeric vector of length p, type index of each feature, allows different variance 
in priors of coefficient of different types}

\item{lassoInit}{TRUE/FALSE, if TRUE, use lasso estimates as MCMC initials}

\item{pi2_weighted}{TRUE/FALSE, if TRUE, the prior of gamma_jk^0 will be weighted by R_j}

\item{pi2_prop_n}{numeric number, specifying n_0 in the proposal distribution of pi_k^0}

\item{MH_ind}{1/0, if 1 MH is used to update pi_k^0; otherwise, Gibbs sampling is used}

\item{burnInIter}{number of iterations as burn-in period}

\item{keepIter}{number of iterations to be stored}

\item{maxIter}{maximum number of iterations}

\item{printInt}{print progress message every print_int iterations}

\item{seed}{integer, specifying the random seed}

\item{alpha2}{numeric number, specifying alpha in the BETA distribution (prior of pi_k^0)}

\item{beta2}{numeric number, specifying beta in the BETA distribution (prior of pi_k^0)}
}
\value{
A list of MCMC output, together with feature selection probability 
and feature Bayesian qvalue
}
\description{
This function applies SOG to binary outcome
}
