% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOG_continuous.R
\name{SOG_continuous}
\alias{SOG_continuous}
\title{SOG function for continuous outcome}
\usage{
SOG_continuous(X, Y, U1, types = NULL, center = FALSE,
  lassoInit = FALSE, BernoulliWeighted = FALSE, pi2_prop_n = 10,
  MH_ind = 1, burnInIter = 1000, keepIter = 3000, maxIter = 10000,
  printInt = 1000, seed = 123, PreEstBeta = FALSE, alpha2 = 1,
  beta2 = 1)
}
\arguments{
\item{X}{n by p feature matrix, where n is the number of samples, 
p is number of features}

\item{Y}{outcome vector of length n}

\item{U1}{p by m1 matirx, denoting group membership of each feature}

\item{types}{numeric vector of length p, type index of each feature, allows different variance 
in priors of coefficient of different types}

\item{center}{TRUE/FALSE, if TRUE, feature feature is centered.}

\item{lassoInit}{TRUE/FALSE, if TRUE, use lasso estimates as MCMC initials}

\item{BernoulliWeighted}{TRUE/FALSE, if TRUE, the prior of gamma_jk^0 will be weighted by R_j}

\item{pi2_prop_n}{numeric number, specifying n_0 in the proposal distribution of pi_k^0}

\item{MH_ind}{1/0, if 1 MH is used to update pi_k^0; otherwise, Gibbs sampling is used}

\item{burnInIter}{number of iterations as burn-in period}

\item{keepIter}{number of iterations to be stored}

\item{maxIter}{maximum number of iterations}

\item{printInt}{print progress message every print_int iterations}

\item{seed}{integer, specifying the random seed}

\item{PreEstBeta}{TRUE/FALSE, if TRUE, use empirical Bayes to pre-estimate alpha and beta in Beta distribution (prior of pi_k^0)}

\item{alpha2}{numeric number, specifying alpha in the BETA distribution (prior of pi_k^0)}

\item{beta2}{numeric number, specifying beta in the BETA distribution (prior of pi_k^0)}
}
\value{
A list of MCMC output, together with feature selection probability 
and feature Bayesian qvalue
}
\description{
This function applies SOG to continous outcome
}
\examples{
\dontrun{
data <- geneSimu1(seed=1)
res <- SOG_continuous(data$X, data$Y, data$U1)
}
}
